orchestrator-service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
httpx==0.24.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-otlp==1.21.0
azure-monitor-opentelemetry-exporter==1.0.0 